refactor: consolidate utils
