+++ config/default.js
@@ -1,4 +1,5 @@
-const { getPackageDetails, isStableVersion } = require('@zap/utils')
+import getPackageDetails from '@zap/utils/getPackageDetails'
+import isStableVersion from '@zap/utils/isStableVersion'
 
 // The current stable base version.
 // If the current version is in the same range asd this, the default database domain will be used.
