+++ electron/menuBuilder.js
@@ -1,9 +1,9 @@
 import os from 'os'
 import { app, Menu, shell, ipcMain } from 'electron'
 import isDev from 'electron-is-dev'
-import { appRootPath } from '@zap/lnd/util'
+import appRootPath from '@zap/utils/appRootPath'
 import { getLanguageName, locales } from '@zap/i18n'
-import { getPackageDetails } from '@zap/utils'
+import getPackageDetails from '@zap/utils/getPackageDetails'
 import openAboutWindow from './about'
 
 const path = require('path')
