+++ renderer/components/Channels/ChannelData.js
@@ -4,7 +4,7 @@ import { FormattedDate, FormattedTime, FormattedMessage } from 'react-intl'
 import styled from 'styled-components'
 import { opacity } from 'styled-system'
 import { Box as BaseBox } from 'rebass'
-import { blockExplorer } from '@zap/utils'
+import blockExplorer from '@zap/utils/blockExplorer'
 import { Bar, DataRow, Link } from 'components/UI'
 import { Truncate } from 'components/Util'
 import { CryptoValue } from 'containers/UI'
