+++ test/unit/utils/validateHost.spec.js
@@ -1,4 +1,4 @@
-import { validateHost } from '@zap/utils/validateHost'
+import validateHost from '@zap/utils/validateHost'
 
 jest.mock('dns')
