+++ utils/connectionString.js
@@ -1,7 +1,7 @@
 import decode from 'lndconnect/decode'
 import parse from 'lndconnect/parse'
 import get from 'lodash.get'
-import { isBase64url } from '@zap/utils'
+import isBase64url from '@zap/utils/isBase64url'
 
 /**
  * Check for a valid lndconnect uri.
