--- utils/index.js
@@ -1,7 +1,0 @@
-export btc from './btc'
-export blockExplorer from './blockExplorer'
-export getPackageDetails from './getPackageDetails'
-export promisifiedCall from './promisifiedCall'
-export isSubDir from './isSubDir'
-export isBase64url from './isBase64url'
-export isStableVersion from './isStableVersion'
